refactor: clear debounce function and cover with tests
