+++ scripts/jest/setupFiles/index.js
@@ -1,1 +1,4 @@
 // Code to setup jest environment
+Object.defineProperty(global, 'performance', {
+	writable: true,
+});
