+++ src/utils/tests.ts
@@ -1,0 +1,1 @@
+export const wait = (time: number) => new Promise((res) => setTimeout(res, time));
